refactor: move JWK types to shared
