+++ packages/shared/src/index.ts
@@ -10,4 +10,5 @@ export * from './ShadowRoot'
 export * from './TypedMessage'
 export * from './utils'
 export * from './hooks'
+export * from './crypto/JWKType'
 export * from './wallet'
